LIBMOBILE-1188
- Initial Readme changes
- updated README.md, settings.gradle.kts, gradle.properties, source package
- deleted MyDestination, MyDestinationTests
- created NielsenDTVRDestination, NielsenDTVRDestinationTests
