+++ gradle.properties
@@ -23,17 +23,17 @@ kotlin.code.style=official
 # Deployment variables
 GROUP=com.segment.analytics.kotlin.destinations
 
-VERSION_CODE=<100>
-VERSION_NAME=<1.0.0>
+VERSION_CODE=100
+VERSION_NAME=1.0.0
 
-POM_ARTIFACT_ID=<destination>
-POM_NAME=<destination>
-POM_DESCRIPTION=<Destination Description>
+POM_ARTIFACT_ID=nielsen-dtvr
+POM_NAME=nielsen-dtvr
+POM_DESCRIPTION=Destination Plugin for Nielsen DTVR
 
-POM_URL=https://github.com/segmentio/<analytics-kotlin-destination>
-POM_SCM_URL=https://github.com/segmentio/<analytics-kotlin-destination>
-POM_SCM_CONNECTION=scm:git:git://github.com/segmentio/<analytics-kotlin-destination>.git
-POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/segmentio/<analytics-kotlin-destination>.git
+POM_URL=https://github.com/segmentio/analytics-kotlin-integration-nielsen-dtvr
+POM_SCM_URL=https://github.com/segmentio/analytics-kotlin-integration-nielsen-dtvr
+POM_SCM_CONNECTION=scm:git:git://github.com/segmentio/analytics-kotlin-integration-nielsen-dtvr.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/segmentio/analytics-kotlin-integration-nielsen-dtvr.git
 
 POM_LICENCE_NAME=The MIT License (MIT)
 POM_LICENCE_URL=http://opensource.org/licenses/MIT
