+++ lib/src/test/kotlin/com/segment/analytics/kotlin/destinations/nielsendtvr/NielsenDTVRDestinationTests.kt
@@ -1,0 +1,10 @@
+package com.segment.analytics.kotlin.destinations.nielsendtvr
+
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+import org.robolectric.annotation.Config
+
+@RunWith(RobolectricTestRunner::class)
+@Config(manifest = Config.NONE)
+class NielsenDTVRDestinationTests {
+}
