+++ settings.gradle.kts
@@ -1,4 +1,4 @@
-rootProject.name = "analytics-kotlin-destination-template"
+rootProject.name = "analytics-kotlin-integration-nielsen-dtvr"
 include(":lib")
 includeBuild("publishing-plugins")
 include(":testapp")
